Move healthcheck helpers to their own file
